std: rename module for clarity
